code/requirements.txt
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.1
Pillow==11.3.0
scikit_learn==1.7.1
seaborn==0.13.2
shap==0.47.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1

